+++ CouponSystem/src/main/java/com/johnbryce/CouponSystem/repo/CustomerRepo.java
@@ -9,8 +9,8 @@ import com.johnbryce.CouponSystem.beans.Coupon;
 import com.johnbryce.CouponSystem.beans.Customer;
 
 @Repository
-public interface CustomerRepo extends JpaRepository<Customer, Integer> {
-	
+public interface CustomerRepo extends JpaRepository<Customer, Long> {
+
 //	public boolean existsByEmailAndPassword(String email, String password);
 //
 //	//public void addCustomer(Customer customer);
